Pulled PaginationControls into its own component
