+++ src/pages/QuestionSection/QuestionSection.module.css
@@ -1,6 +1,3 @@
-.paginationContainer {
-  margin-bottom: 1rem;
-}
 .paginationLinks {
   display: flex;
   justify-content: space-evenly;
