+++ src/pages/QuestionSection/components/PaginationControls/PaginationControls.module.css
@@ -1,0 +1,7 @@
+.container {
+  margin-bottom: 1rem;
+}
+.links {
+  display: flex;
+  justify-content: space-evenly;
+}
